出入库通知单接收
出入库通知单接收

+++ src/main/java/com/rfid/platform/entity/TagStorageOperationBean.java
@@ -21,7 +21,7 @@ public class TagStorageOperationBean implements Serializable {
     private String noticeNo;
 
     @TableField(value = "NOTICE_QUANTITY")
-    private Integer noticeQuantity;
+    private Long noticeQuantity;
 
     @TableField(value = "NOTICE_TYPE")
     private Integer noticeType;
